Generic resource manager
A templated resource manager that returns a shared_ptr to the managed resource. Can request the manager to relinquish ownership and it will just maintain a weak pointer to the object. A later request to obtain the object will result in an allocation only if the weak pointer is pointing to an expired object. This manager is meant to manage really large memory consuming resources that take a while to load.